requirements.txt

streamlit
pandas
plotly
requests
prophet
streamlit-autorefresh
mysql.connector


